Add monitoring for new channels (#6)
* Add monitoring for new channels

* Fix monitoring for new channels

Fixed 5 minor issues with pull request. These issues were
highlighted in the comments to the pull request. The only
issue that wasn't resolved was:

func writeChannelsToColl(coll *mgo.Collection, chans []ChannelAndID) error {
	coll.DropCollection()
	for _, v := range chans {
		if err := coll.Insert(v); err != nil {

The for loop was left in because Go will not convert a
[]T []Interface{} for you
https://stackoverflow.com/a/12754757/859353
(as pointed out by J.W.)

* Remove some whitespace

* Simplify some error handling